Move string <-> filename algorithm to separate module
