+++ src/fontra/backends/fontra.py
@@ -4,12 +4,13 @@ import pathlib
 import shutil
 from copy import deepcopy
 from dataclasses import asdict
-from urllib.parse import unquote
 
 import dacite
 
 from fontra.core.classes import Font, VariableGlyph
 from fontra.core.path import PackedPath, Path
+
+from .filenames import stringToFileName
 
 FILENAME_GLYPH_INFO = "glyph-info.csv"
 FILENAME_FONT_DATA = "font-data.json"
@@ -158,75 +159,3 @@ daciteConfig = dacite.Config(
     type_hooks={PackedPath: _ensurePackedPathData, Path: _ensurePathData},
     strict_unions_match=False,
 )
-
-
-#
-# Glyph name to file name algorithm, originally proposed for UFO
-# See also https://github.com/unified-font-object/ufo-spec/issues/164
-#
-
-
-separatorChar = "^"
-
-# TODO: [insert references]
-reservedCharacters = set(' " % * + / : < > ? [ \\ ] | '.split())
-reservedCharacters.update(chr(i) for i in range(32))
-# reservedCharacters.add(" ")  # should we escape space chars or not?
-reservedCharacters.add(chr(0x7F))
-reservedCharacters.add(separatorChar)
-assert all(len(c) == 1 for c in reservedCharacters)
-
-
-# TODO: [insert references]
-reservedFileNames = set(
-    """
-CON
-PRN
-AUX
-CLOCK$
-NUL
-COM1
-LPT1
-LPT2
-LPT3
-COM2
-COM3
-COM4
-""".lower().split()
-)
-
-
-base32chars = "0123456789ABCDEFGHIJKLMNOPQRSTUV"
-assert len(set(base32chars)) == 32
-
-
-def stringToFileName(string):
-    codeDigits = []
-    for i in range(0, len(string), 5):
-        digit = 0
-        bit = 1
-        for c in string[i : i + 5]:
-            if c.isupper():
-                digit |= bit
-            bit <<= 1
-        codeDigits.append(digit)
-    # strip trailing zeros
-    while codeDigits and codeDigits[-1] == 0:
-        codeDigits.pop()
-    fileName = "".join(
-        f"%{ord(c):02X}" if c in reservedCharacters else c for c in string
-    )
-    if fileName[0] == ".":
-        fileName = "%2E" + fileName[1:]
-    if not codeDigits and fileName.lower() in reservedFileNames:
-        codeDigits = [0]
-    if codeDigits:
-        disambiguationCode = separatorChar + "".join(base32chars[d] for d in codeDigits)
-    else:
-        disambiguationCode = ""
-    return fileName + disambiguationCode
-
-
-def fileNameToString(fileName):
-    string = fileName.split(separatorChar, 1)[0]
-    return unquote(string, encoding="ascii", errors="strict")
